[General] Added GameSettings
Trying to implement CAO in a game agnostic way was an error. Each game
has its specificities that require specific code changes. This is
probably just the beginning if I want to fully support each game
This does not mean custom games won't be supported, a custom profile
could be read from JSON